Factor out  function to utils
