+++ frontend/src/utils.js
@@ -64,6 +64,16 @@ export const getTypeIcon = (type) => {
   return mapping[type] || null
 }
 
+export const getType = (type) => {
+  const mapping = {
+    plant: "Plante",
+    microorganism: "Micro-organisme",
+    ingredient: "Ingredient",
+    substance: "Substance",
+  }
+  return mapping[type] || null
+}
+
 export const headers = {
   "X-CSRFToken": window.CSRF_TOKEN || "",
   "Content-Type": "application/json",
